Enable writing data from Hive map-joins (ie formatter by the Lazy SerDe)

+++ connector/src/main/java/com/google/cloud/hive/bigquery/connector/utils/avro/AvroUtils.java
@@ -117,7 +117,7 @@ public class AvroUtils {
       return schema;
     }
     if (fieldOi instanceof FloatObjectInspector) {
-      return Schema.create(Schema.Type.FLOAT);
+      return Schema.create(Schema.Type.DOUBLE);
     }
     if (fieldOi instanceof DoubleObjectInspector) {
       return Schema.create(Schema.Type.DOUBLE);
